# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: S.A.L.I.M. Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── USERS ──────────────────────────────────────────────
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── OBJECTIVES ─────────────────────────────────────────
...
  /habits:
    get:
      operationId: listHabits
      tags: [habits]
      summary: List habits for current user
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [habits]
      summary: Create a new habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitInput"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /habits/{id}:
    patch:
      operationId: updateHabit
      tags: [habits]
      summary: Update a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitUpdate"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
    delete:
      operationId: deleteHabit
      tags: [habits]
      summary: Delete a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /habits/{id}/check:
    post:
      operationId: checkHabit
      tags: [habits]
      summary: Log a habit completion for today
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitLog"

  # ─── JOURNAL ────────────────────────────────────────────
  /journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List journal entries
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create a journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /journal/{id}:
    patch:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DASHBOARD ──────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── AUTH ──────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, plan, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        plan:
          type: string
          enum: [free, premium]
        mission:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        mission:
          type: string
        avatarUrl:
          type: string

    # ─── OBJECTIVES ────────────────────────────────────────
...
    Habit:
      type: object
      required: [id, userId, title, frequency, currentStreak, longestStreak, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        objectiveId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        frequency:
          type: string
          enum: [daily, weekdays, weekly]
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        totalCompletions:
          type: integer
        isActive:
          type: boolean
        lastCheckedDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
    HabitInput:
      type: object
      required: [title, frequency]
      properties:
        objectiveId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekdays, weekly]
    HabitUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekdays, weekly]
        isActive:
          type: boolean
    HabitLog:
      type: object
      required: [id, habitId, date]
      properties:
        id:
          type: integer
        habitId:
          type: integer
        date:
          type: string
          format: date
        completedAt:
          type: string
          format: date-time

    # ─── JOURNAL ───────────────────────────────────────────
    JournalEntry:
      type: object
      required: [id, userId, content, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        objectiveId:
          type: ["integer", "null"]
        title:
          type: ["string", "null"]
        content:
          type: string
        mood:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    JournalInput:
      type: object
      required: [content]
      properties:
        objectiveId:
          type: integer
        title:
          type: string
        content:
          type: string
          minLength: 1
        mood:
          type: integer
        tags:
          type: array
          items:
            type: string
    JournalUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
          minLength: 1
        mood:
          type: integer
        tags:
          type: array
          items:
            type: string

    # ─── DASHBOARD ─────────────────────────────────────────
...