# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * S.A.L.I.M. Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from "./userPlan";

export interface User {
  id: number;
  email: string;
  name: string;
  plan: UserPlan;
  /** @nullable */
  mission?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: Date;
}